backend/requirements.txt
fastapi
uvicorn
azure-storage-blob
azure-cosmos
python-multipart
pydantic
python-dotenv